♻️ Refactor: 로그아웃/회원탈퇴 호출 뷰모델 변경
- AuthViewModel -> SettingViewModel
- 하드 코딩 수정

Related to: #357
